refactor(controller): 패키지 구조 정리로 controller 위치 변경
- 클래스 로직 변경 없이 import 경로 및 Config 의존성만 수정
