+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.config.LottoConfig;
-import lotto.domain.controller.LottoController;
+import lotto.controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
