+++ src/test/java/lotto/domain/LottoControllerTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import lotto.config.LottoConfig;
-import lotto.domain.controller.LottoController;
+import lotto.controller.LottoController;
 import org.junit.jupiter.api.Test;
 
 public class LottoControllerTest extends NsTest {
